pyproject.toml
[tool.poetry]
name = "smah"
version = "0.1.4"
description = ""
authors = ["Keith Brings <keith.brings@noizu.com>"]
readme = "README.md"
package-mode = true

[tool.poetry.dependencies]
python = "^3.10"
rich = "^13.9.2"
openai = "^1.52.0"
aiohttp = "^3.10.10"
requests = "^2.32.3"
pyyaml = "^6.0.2"
readchar = "^4.2.0"
prompt-toolkit = "^3.0.48"
psutil = "^6.1.0"

[tool.poetry.scripts]
smah = "smah.smah:main"
smah-db = "smah.smah_migrate:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
